Initial "0_led_toggle_addr" program.
This code follows along with Section 1: Videos 5-7.

I have made slight changes to the code (which was written for the NUCLEO-F411RE) so that it functions with the NUCLEO-F412ZG board.

The code was designed to use LED "LD1", which is typically assigned to PA5. However, LD1 is assigned elsewhere on the F412ZG board. So, I decided to use LD2, which is assigned to PB7. I have added additional constants and commented out the original code in the main method.